data/requirements.txt
# --- Config & Utilities ---
python-dotenv==1.2.1
boto3
requests==2.32.5

# --- Pipeline (EDA/Ingest) ---
pandas==2.3.3
pypdf==6.1.3
opensearch-py==2.8.0

# --- RAGAS & Tests ---
# Ragas es la dependencia principal que fijará las versiones de langchain
ragas==0.1.21
datasets==2.18.0
pytest==8.4.2

# --- Embeddings (HF, Transformers) ---
# Dejamos que pip encuentre una versión compatible
langchain-huggingface 
sentence-transformers==5.1.2